Moving vtctld core into go/vt/vtctld.
