+++ go/cmd/vtctld/plugin_etcdtopo.go
@@ -10,13 +10,14 @@ import (
 	"github.com/youtube/vitess/go/vt/etcdtopo"
 	"github.com/youtube/vitess/go/vt/servenv"
 	"github.com/youtube/vitess/go/vt/topo"
+	"github.com/youtube/vitess/go/vt/vtctld"
 )
 
 func init() {
 	// Wait until flags are parsed, so we can check which topo server is in use.
 	servenv.OnRun(func() {
 		if etcdServer, ok := topo.GetServer().Impl.(*etcdtopo.Server); ok {
-			HandleExplorer("etcd", etcdtopo.NewExplorer(etcdServer))
+			vtctld.HandleExplorer("etcd", etcdtopo.NewExplorer(etcdServer))
 		}
 	})
 }
