+++ go/cmd/vtctld/plugin_grpcvtctlserver.go
@@ -1,4 +1,6 @@
-// Copyright 2015, Google Inc. All rights reserved.                                                                                                                                                   // Use of this source code is governed by a BSD-style                                                                                                                                                 // license that can be found in the LICENSE file.
+// Copyright 2015, Google Inc. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 
 package main
