+++ go/cmd/vtctld/plugin_zktopo.go
@@ -9,6 +9,7 @@ package main
 import (
 	"github.com/youtube/vitess/go/vt/servenv"
 	"github.com/youtube/vitess/go/vt/topo"
+	"github.com/youtube/vitess/go/vt/vtctld"
 	"github.com/youtube/vitess/go/vt/zktopo"
 )
 
@@ -16,7 +17,7 @@ func init() {
 	// Wait until flags are parsed, so we can check which topo server is in use.
 	servenv.OnRun(func() {
 		if zkServer, ok := topo.GetServer().Impl.(*zktopo.Server); ok {
-			HandleExplorer("zk", zktopo.NewZkExplorer(zkServer.GetZConn()))
+			vtctld.HandleExplorer("zk", zktopo.NewZkExplorer(zkServer.GetZConn()))
 		}
 	})
 }
