+++ go/vt/vtctld/action_repository.go
@@ -1,4 +1,4 @@
-package main
+package vtctld
 
 import (
 	"flag"
