+++ go/vt/vtctld/explorer_test.go
@@ -1,8 +1,4 @@
-// Copyright 2015, Google Inc. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package main
+package vtctld
 
 import (
 	"net/http"
