+++ go/vt/vtctld/tablet_data.go
@@ -1,4 +1,4 @@
-package main
+package vtctld
 
 import (
 	"flag"
