+++ go/vt/vtctld/tablet_data_test.go
@@ -1,4 +1,4 @@
-package main
+package vtctld
 
 import (
 	"sync"
